Preference change notification rewrite.
The main change here is the rewrite of the preferences change
stuff. To handle this, the GConfHandler class was written. It makes
gconf watch some variables and call a callback when each property is
changed.

Another thing done is to implement all features added to the
preferences windo. (Closes: #29, #62, #70, #88)

Bot be carefull, some working things were broken, like keyboard
shortcut edit and compatibility keyboard things. I'm working on it and
in the next commit I think every thing will be working again.