Making links clickable (Closes: #38).
Now links are highlighted by vte and the button-press-event is being
handled by a callback that looks for a matched string and if it
exists, gnome-open or xdg-open will be spawned to decide who should
handle the found uri. I implemented it with a spawn to some of that
programs because gtk_show_uri is not binded to python yet.