Fix dual monitor problem.
Fix the dual monitor problem that manifested with guake appearing on the
monitor with the width of the other monitor. The problem was that guake
assumed that left monitor is always the default monitor, which is not
always the case, and always appeared on position 0,0 (top left of the
left monitor). In my case it took the width of the monitor0 1680 but
appeared on the left monitor which has width of 1280. To fix this,
rework the get_final_window_size to get_final_window_rect that will
return the rectangle (x,y,width,height) of the window. Also remove some
ilogical code that compared the monitor height with height preference
(height preference is in percents and monitor height is in pixels).