Adding select_tab and get_selected_tab.
Now it is possible to get and set the selected tab
in guake through command line and dbus.
Another important thing done in this commit is
that the short option to --new-tab was changed
to -n instead of -t and the --show-hide option
was changed by --toggle-visibility and its short
to -t.