don't rip out underscores from terminal titles
As radiobutton labels are now set to terminal titles, using underscores in the
labels to create accelerators becomes harmful. In most cases the underscore will
be a part of dirname or command or file name, so user wouldn't want that
accelerator, and will instead be confused of underscore just disappearing.

Anyway, we should get some more usable accelerator scheme for the tabs
(like in #233)