Hides tab bar completely when opening tab in fullscreen.
When in fullscreen, creating a new tab will always cause a part
of the tab bar to appear. This solution works, but is a hack.
Someone more familiar with the code definitely should be able
to come up with something better.

Steps to reproduce:
1. Disable option 'show tab bar'
2. Start guake
3. Go into fullscreen
4. Open a new tab

Happened on ubuntu 12.04--13.10 on all systems I've used,
including mid 2012 macbook pro and vmware workstation 9.