Fix #619 : Unable to set default interpreter
This pull request fixes https://github.com/Guake/guake/issues/619. The issue is similar to https://github.com/Guake/guake/issues/578 where the data in 'prefs.glade' has not been linked to the appropriate handler. In a similar manner the 'use as login shell' check-box is also not linked correctly.