Added support for selected text in custom commands
Now it's possible to pass the in terminal selected text to custom commands, to do that just put the number 1 in the command array withing the custom command file.
For example the following json entry launches the command "stat #filename#" and filename is the terminal selection:

...
{
        "description": "File stat",
        "cmd": [
                    "stat",
                    1
                ]
},
...

Note that the '1' flag must be and integer and not a string, so avoid enclosing in double quotes.
Other numbers could be used for future features.
If there is not a selection the command "stat" is issued with no parameters.