Implement option for 8-bit bright text
In vte 0.52, a new option was added to emit only bold when bold text is
emitted, and made the default behavior in vte 0.56. For details, see
https://bugzilla.gnome.org/show_bug.cgi?id=762247

The motivation for the change is to allow better 256-bit color in applications
and colorschemes. However, traditionally, applications assumed that bold text
would also be bright, and many applications look quite bad with this setting.
It is therefore useful to allow users to decide whether to utilize this
behavior.

Fixes #1507
Closes #1089