Block notebook button-press-event when search box visible
If we don't block notebook button-press-event, it will affect search box,
when double click on search box button or search entry, it will then add
a new page with focus, which is what we don't want.

So, block this event when search box is visible, and unblock it when
search box is show off.