use importlib_metadata rather than pbr to introspect versions at runtime
Use of pbr at runtime is super ugly, as all it does is use pkg_resources to
parse distribution versions and pkg_resources could do the exact same thing. But
they are both slow, because pkg_resources is (by design) slow.

importlib.metadata was added to the stdlib in python 3.8 to fulfill this use
case in a speedy and elegant manner.

FIXME: this entire schema is ugly, not merely at runtime -- use of pbr for
versioning from git is an outdated and stale idea, this should use
setuptools_scm instead which:
- is standard
- supports writing the version to _version.py