remove useless import of pkg_resources
pkg_resources is not guaranteed to be installed, since it's not in
install_requires. pbr does depend on it, but if pbr was dropped from
runtime then pkg_resources would not be available.

Furthermore, pbr upstream doesn't add a dependency on pkg_resources, even
though the code relies on it. So unless your pbr module comes from a well
maintained linux distro, it's perfectly possible to install pbr with pip
and not have pkg_resources.

It's all pointless anyway, since guake does not nor ever has used
pkg_resources and should never have imported it without using it. So we'll
just stop importing it.