remove broken and unused codepath
This method is never invoked; if it was, the fact that it takes zero
arguments, not even self, and then tries to access self, would prevent it
from working.

It was added in commit bdab3af5ef14baf22dae147d191f8187c4567922 but the
feature works fine even without it.