setup.cfg: directly specify needed install requirements
Do not rely on running pip to install dependencies, which e.g. distros cannot
do. Instead let this be specified directly as package metadata.

While we are at it, stop actually running pip install if DESTDIR is set, since
this is a fairly likely sign that guake is being installed in a distro
environment where pip install is not allowed.