Attempt to continue if schema file is in a protected location
While purging my environment and reinstalling, I realised that if the schema file is in a protected location, this still fails on the first run because the version doesn't match and it'll try to recompile and fail. This probably will happen with packaged versions that put the schema in a protected place too. sudo make install and package managers should be compiling the file properly for us before this, so we can go ahead and keep going.