setup: replace the last remnant of pbr with setuptools_scm
setuptools_scm is a standard and robust tool for dynamically getting
version numbers from e.g. git; among its features is the ability to
write the results to a python module that the application can then
import from. This removes the overhead of large modules which introspect
the dist-info to get the version, in favor of importing a frozen string.