Enable the use of AyatanaAppIndicator
This uses the newer successor to AppIndicator3 for the tray icon implementation, with a fallback to the older AppIndicator3 if the Ayatana version is not found. AyatanaAppIndicator is backwards compatible with AppIndicator3, so the API remains the same. The icon is also fixed.

Ref https://github.com/BuddiesOfBudgie/budgie-desktop/issues/446
Fixes https://github.com/Guake/guake/issues/433

sem-ver: bug-fix

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>